Remove network proxy token cache (#459)
When firefox is restarted, after a session with the proxy on, the network
component doesn't have the proxy token yet and because of this, the first
requests will failed.
This commit makes the network component to ask for the token directly to fxa.js.
Because of this, it will always have a valid token, if it exists. The cache is
not needed anymore.